# backend/swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Real-time management system for localization strings.",
    "title": "LocalizeMe.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Return all groups.",
        "operationId": "FindAllGroups",
        "responses": {
          "200": {
            "description": "Group",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Group"
              }
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "500": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Groups"
        ],
        "summary": "Update the information of a group.",
        "operationId": "UpdateGroup",
        "parameters": [
          {
            "name": "Group",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "422": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Groups"
        ],
        "summary": "Create a new group.",
        "operationId": "CreateGroup",
        "parameters": [
          {
            "name": "GroupDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "422": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete a group by id.",
        "operationId": "DeleteGroup",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "description": "The group´s id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/bool"
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Groups"
        ],
        "summary": "Disable of a group.",
        "operationId": "DisableGroup",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "description": "The group´s id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Group",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      }
    },
    "/language": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Languages"
        ],
        "summary": "Update the information of a language.",
        "operationId": "UpdateLanguage",
        "parameters": [
          {
            "name": "Language",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Language",
            "schema": {
              "$ref": "#/definitions/Language"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "422": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      }
    },
    "/languages": {
      "get": {
        "tags": [
          "Languages"
        ],
        "summary": "Return all languages.",
        "operationId": "FindAllLanguages",
        "responses": {
          "200": {
            "description": "Language",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Language"
              }
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Languages"
        ],
        "summary": "Create a new language.",
        "operationId": "CreateLanguage",
        "parameters": [
          {
            "name": "LanguageDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LanguageDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Language",
            "schema": {
              "$ref": "#/definitions/Language"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "422": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      }
    },
    "/languages/{id}": {
      "delete": {
        "tags": [
          "Languages"
        ],
        "summary": "Return a language by id.",
        "operationId": "DeleteLanguage",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "description": "The language´s id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/bool"
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Languages"
        ],
        "summary": "Disable of a language.",
        "operationId": "DisableLanguage",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "description": "The language´s id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Language",
            "schema": {
              "$ref": "#/definitions/Language"
            }
          },
          "400": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "401": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          },
          "403": {
            "description": "ErrorDto",
            "schema": {
              "$ref": "#/definitions/ErrorDto"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorDto": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain/dto"
    },
...
    "Group": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "$ref": "#/definitions/User"
        },
        "permissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Permission"
          },
          "x-go-name": "Permissions"
        },
        "public": {
          "type": "boolean",
          "x-go-name": "Public"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain"
    },
    "GroupDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "$ref": "#/definitions/User"
        },
        "permissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Permission"
          },
          "x-go-name": "Permissions"
        },
        "public": {
          "type": "boolean",
          "x-go-name": "Public"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain/dto"
    },
    "Language": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "isoCode": {
          "type": "string",
          "x-go-name": "IsoCode"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain"
    },
    "LanguageDto": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "isoCode": {
          "type": "string",
          "x-go-name": "IsoCode"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain/dto"
    },
    "ObjectID": {
      "type": "array",
      "title": "ObjectID is the BSON ObjectID type.",
      "items": {
        "type": "integer",
        "format": "uint8"
      },
      "x-go-package": "go.mongodb.org/mongo-driver/bson/primitive"
    },
    "Permission": {
      "type": "object",
      "properties": {
        "canWrite": {
          "type": "boolean",
          "x-go-name": "CanWrite"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain"
    },
...
    "User": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "admin": {
          "type": "boolean",
          "x-go-name": "Admin"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "gitlab.com/HP-SCDS/Observatorio/2021-2022/localizeme/uniovi-localizeme/internal/core/domain"
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [
        "[]"
      ]
    }
  ]
}